# Refactor `CallbackStack` based on new "pack locking" strategy