Fixes https://github.com/typelevel/cats-effect/issues/3940. Very fun collaboration with @samspills but credit to @mernst-github for proposing the idea of guarding `pack` in https://github.com/typelevel/cats-effect/issues/3940#issuecomment-1883123013.

The new callback stack is made up of two data types:
1. `CallbackStack.Node`, a singly-linked list of callbacks
2. `CallbackStack` itself, which holds:
    - an `AtomicReference` to the head `Node`, and
    - an `AtomicBoolean` lock `allowedToPack`, used for guarding `pack`

Pushing to the stack is a straighforward CAS of the head.

`pack` is the only method that modifies the list.
- the `AtomicBoolean` lock guarantees that only one `pack` operation is running at a time, and is also used to publish the modifications to subsequent packs
- if a call to `pack` is unable to acquire the lock, it gives up immediately

Invoking the callbacks is not guarded by the lock. This means that the modifications by `pack` may not be visible, and in fact `pack`ing can even occur concurrently while iterating over the list. This is actually okay:
- the list is only ever modified to _remove_ `Node`s, since additions happen by CASing the pointer to the head `Node`
- thus, if while iterating over the list we follow a stale pointer to a next `Node` that was actually removed, we take a "detour" but do eventually visit all the `Node`s in the list