@djspiewak 2024-01-14
Looks like a perfect wash (within margin for error) in terms of performance. Good to go on that front!

## Before

```
[info] Benchmark                                             (count)  (cpuTokens)   (size)   Mode  Cnt      Score     Error    Units
[info] DeepBindBenchmark.async                                   N/A          N/A    10000  thrpt   10   2935.967 ± 102.359    ops/s
[info] DeepBindBenchmark.delay                                   N/A          N/A    10000  thrpt   10   9742.384 ±  42.545    ops/s
[info] DeepBindBenchmark.pure                                    N/A          N/A    10000  thrpt   10  11626.005 ± 170.858    ops/s
[info] DeferredBenchmark.cancel                                   10          N/A      N/A  thrpt   10    104.204 ±   0.456    ops/s
[info] DeferredBenchmark.cancel                                  100          N/A      N/A  thrpt   10     13.537 ±   0.074    ops/s
[info] DeferredBenchmark.cancel                                 1000          N/A      N/A  thrpt   10      1.312 ±   0.008    ops/s
[info] DeferredBenchmark.complete                                 10          N/A      N/A  thrpt   10    110.424 ±   0.836    ops/s
[info] DeferredBenchmark.complete                                100          N/A      N/A  thrpt   10     26.084 ±   0.109    ops/s
[info] DeferredBenchmark.complete                               1000          N/A      N/A  thrpt   10      5.073 ±   0.287    ops/s
[info] DeferredBenchmark.getAfter                                 10          N/A      N/A  thrpt   10   7059.728 ±  19.859    ops/s
[info] DeferredBenchmark.getAfter                                100          N/A      N/A  thrpt   10   2033.584 ±  10.383    ops/s
[info] DeferredBenchmark.getAfter                               1000          N/A      N/A  thrpt   10    247.342 ±   0.501    ops/s
[info] DeferredBenchmark.getBefore                                10          N/A      N/A  thrpt   10   3994.309 ±  15.618    ops/s
[info] DeferredBenchmark.getBefore                               100          N/A      N/A  thrpt   10    798.325 ±   1.832    ops/s
[info] DeferredBenchmark.getBefore                              1000          N/A      N/A  thrpt   10     81.027 ±   0.163    ops/s
[info] ParallelBenchmark.parTraverse                             N/A        10000     1000  thrpt   10    878.708 ±  25.632    ops/s
[info] ParallelBenchmark.traverse                                N/A        10000     1000  thrpt   10     70.062 ±   0.050    ops/s
[info] ShallowBindBenchmark.async                                N/A          N/A    10000  thrpt   10   2157.575 ±   3.120    ops/s
[info] ShallowBindBenchmark.delay                                N/A          N/A    10000  thrpt   10  10052.708 ±  54.888    ops/s
[info] ShallowBindBenchmark.pure                                 N/A          N/A    10000  thrpt   10  10539.875 ± 135.470    ops/s
[info] WorkStealingBenchmark.alloc                               N/A          N/A  1000000  thrpt   10     14.151 ±   0.111  ops/min
[info] WorkStealingBenchmark.manyThreadsSchedulingBenchmark      N/A          N/A  1000000  thrpt   10     50.544 ±   0.753  ops/min
[info] WorkStealingBenchmark.runnableScheduling                  N/A          N/A  1000000  thrpt   10   2958.068 ±  15.379  ops/min
[info] WorkStealingBenchmark.runnableSchedulingScalaGlobal       N/A          N/A  1000000  thrpt   10   2195.662 ±   5.718  ops/min
[info] WorkStealingBenchmark.scheduling                          N/A          N/A  1000000  thrpt   10     53.854 ±   2.195  ops/min
```

## After

```
[info] Benchmark                                             (count)  (cpuTokens)   (size)   Mode  Cnt      Score     Error    Units
[info] DeepBindBenchmark.async                                   N/A          N/A    10000  thrpt   10   3068.624 ±  16.441    ops/s
[info] DeepBindBenchmark.delay                                   N/A          N/A    10000  thrpt   10   9853.126 ± 124.913    ops/s
[info] DeepBindBenchmark.pure                                    N/A          N/A    10000  thrpt   10  11664.990 ±  94.674    ops/s
[info] DeferredBenchmark.cancel                                   10          N/A      N/A  thrpt   10    107.015 ±   0.686    ops/s
[info] DeferredBenchmark.cancel                                  100          N/A      N/A  thrpt   10     14.352 ±   0.083    ops/s
[info] DeferredBenchmark.cancel                                 1000          N/A      N/A  thrpt   10      1.326 ±   0.016    ops/s
[info] DeferredBenchmark.complete                                 10          N/A      N/A  thrpt   10    110.982 ±   1.658    ops/s
[info] DeferredBenchmark.complete                                100          N/A      N/A  thrpt   10     26.111 ±   0.183    ops/s
[info] DeferredBenchmark.complete                               1000          N/A      N/A  thrpt   10      5.129 ±   0.271    ops/s
[info] DeferredBenchmark.getAfter                                 10          N/A      N/A  thrpt   10   6990.249 ±  47.301    ops/s
[info] DeferredBenchmark.getAfter                                100          N/A      N/A  thrpt   10   2042.712 ±  21.909    ops/s
[info] DeferredBenchmark.getAfter                               1000          N/A      N/A  thrpt   10    246.581 ±   1.930    ops/s
[info] DeferredBenchmark.getBefore                                10          N/A      N/A  thrpt   10   3941.678 ±  24.476    ops/s
[info] DeferredBenchmark.getBefore                               100          N/A      N/A  thrpt   10    744.944 ±   5.293    ops/s
[info] DeferredBenchmark.getBefore                              1000          N/A      N/A  thrpt   10     80.747 ±   0.431    ops/s
[info] ParallelBenchmark.parTraverse                             N/A        10000     1000  thrpt   10    887.059 ±  23.448    ops/s
[info] ParallelBenchmark.traverse                                N/A        10000     1000  thrpt   10     69.816 ±   0.752    ops/s
[info] ShallowBindBenchmark.async                                N/A          N/A    10000  thrpt   10   2169.835 ±  22.053    ops/s
[info] ShallowBindBenchmark.delay                                N/A          N/A    10000  thrpt   10   9872.628 ± 116.515    ops/s
[info] ShallowBindBenchmark.pure                                 N/A          N/A    10000  thrpt   10  10410.052 ±  75.004    ops/s
[info] WorkStealingBenchmark.alloc                               N/A          N/A  1000000  thrpt   10     14.018 ±   0.133  ops/min
[info] WorkStealingBenchmark.manyThreadsSchedulingBenchmark      N/A          N/A  1000000  thrpt   10     48.573 ±   2.040  ops/min
[info] WorkStealingBenchmark.runnableScheduling                  N/A          N/A  1000000  thrpt   10   2997.308 ±  10.291  ops/min
[info] WorkStealingBenchmark.runnableSchedulingScalaGlobal       N/A          N/A  1000000  thrpt   10   2239.119 ±  12.325  ops/min
[info] WorkStealingBenchmark.scheduling                          N/A          N/A  1000000  thrpt   10     52.694 ±   1.716  ops/min
```